requirements.txt
APScheduler==3.6.3
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
tzlocal==2.0.0
Werkzeug==1.0.0
WTForms==2.2.1
